+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.32 branch
 ### 2.32.1
+* the `UsersTable` belongs to `UsersGroups` and no longer to `Groups` (the name of the association has changed). This
+  avoids using a reserved word for mysql. The property name (`group`) remained unchanged;
 * improved the query filter for posts and pages in the admin panel: now the `title` field works for both title and slug;
 * `User::_getLastLogins()` returns an array and no longer a collection;
 * `assertSqlEndsWith()` and `assertSqlEndsNotWith()` methods have been replaced with `assertStringEndsWith()` and